fix(recap): Requested changes for unextracted_rds command
